Refactoring and dependencies change in gobblin-compliance (#1798)
