+++ gobblin-modules/gobblin-compliance/build.gradle
@@ -19,7 +19,8 @@ apply plugin: 'java'
 
 dependencies {
   compile project(":gobblin-data-management")
-  compile project(":gobblin-modules:gobblin-azkaban")
+  compile externalDependency.azkaban
+  compile externalDependency.hiveJdbc
 }
 
 test {
